requirements.txt
Django>=4.0,<5.0
djangorestframework
djoser
drf-yasg
django-celery-beat
celery
redis
django-redis
